Replaced %<n>.<n>d with %0<n>d

+++ src/gmt_calclock.c
@@ -909,23 +909,23 @@ void GMT_get_time_label (struct GMT_CTRL *C, char *string, struct GMT_PLOT_CALCL
 
 	switch (T->unit) {
 		case 'Y':	/* 4-digit integer year */
-			(P->date.compact) ? sprintf (string, "%d", calendar.year) : sprintf (string, "%4.4d", calendar.year);
+			(P->date.compact) ? sprintf (string, "%d", calendar.year) : sprintf (string, "%04d", calendar.year);
 			break;
 		case 'y':	/* 2-digit integer year */
-			/* (P->date.compact) ? sprintf (string, "%d", calendar.year % 100) : sprintf (string, "%2.2d", calendar.year % 100); */
-			sprintf (string, "%2.2d", calendar.year % 100);
+			/* (P->date.compact) ? sprintf (string, "%d", calendar.year % 100) : sprintf (string, "%02d", calendar.year % 100); */
+			sprintf (string, "%02d", calendar.year % 100);
 			break;
 		case 'O':	/* Plot via date format */
 			GMT_format_calendar (C, string, NULL, &P->date, &P->clock, T->upper_case, T->flavor, t);
 			break;
 		case 'o':	/* 2-digit month */
-			(P->date.compact) ? sprintf (string, "%d", calendar.month) : sprintf (string, "%2.2d", calendar.month);
+			(P->date.compact) ? sprintf (string, "%d", calendar.month) : sprintf (string, "%02d", calendar.month);
 			break;
 		case 'U':	/* ISO year, week, day via date format */
 			GMT_format_calendar (C, string, NULL, &P->date, &P->clock, T->upper_case, T->flavor, t);
 			break;
 		case 'u':	/* 2-digit ISO week */		
-			(P->date.compact) ? sprintf (string, "%d", calendar.iso_w) : sprintf (string, "%2.2d", calendar.iso_w);
+			(P->date.compact) ? sprintf (string, "%d", calendar.iso_w) : sprintf (string, "%02d", calendar.iso_w);
 			break;
 		case 'K':	/*  ISO Weekday name */
 			if (T->upper_case) GMT_str_toupper (C->current.time.language.day_name[T->flavor][calendar.iso_d%7]);
@@ -940,21 +940,21 @@ void GMT_get_time_label (struct GMT_CTRL *C, char *string, struct GMT_PLOT_CALCL
 		case 'd':	/* 2-digit day or 3-digit day of year */
 		case 'R':	/* Gregorian month-days are the same thing - only they start at beginning of weeks and not months */
 			if (P->date.day_of_year)
-				(P->date.compact) ? sprintf (string, "%d", calendar.day_y) : sprintf (string, "%3.3d", calendar.day_y);
+				(P->date.compact) ? sprintf (string, "%d", calendar.day_y) : sprintf (string, "%03d", calendar.day_y);
 			else
-				(P->date.compact) ? sprintf (string, "%d", calendar.day_m) : sprintf (string, "%2.2d", calendar.day_m);
+				(P->date.compact) ? sprintf (string, "%d", calendar.day_m) : sprintf (string, "%02d", calendar.day_m);
 			break;
 		case 'H':	/* Hours via clock format */
 			GMT_format_calendar (C, NULL, string, &P->date, &P->clock, T->upper_case, T->flavor, t);
 			break;
 		case 'h':	/* 2-digit hour */
-			(P->date.compact) ? sprintf (string, "%d", calendar.hour) : sprintf (string, "%2.2d", calendar.hour);
+			(P->date.compact) ? sprintf (string, "%d", calendar.hour) : sprintf (string, "%02d", calendar.hour);
 			break;
 		case 'M':	/* Minutes via clock format */
 			GMT_format_calendar (C, NULL, string, &P->date, &P->clock, T->upper_case, T->flavor, t);
 			break;
 		case 'm':	/* 2-digit minutes */
-			(P->date.compact) ? sprintf (string, "%d", calendar.min) : sprintf (string, "%2.2d", calendar.min);
+			(P->date.compact) ? sprintf (string, "%d", calendar.min) : sprintf (string, "%02d", calendar.min);
 			break;
 #ifdef GMT_COMPAT
 		case 'C':
@@ -968,7 +968,7 @@ void GMT_get_time_label (struct GMT_CTRL *C, char *string, struct GMT_PLOT_CALCL
 			GMT_report (C, GMT_MSG_COMPAT, "Warning: Unit c for seconds is deprecated; use s.\n");
 #endif
 		case 's':	/* 2-digit seconds */
-			(P->date.compact) ? sprintf (string, "%ld", lrint(calendar.sec)) : sprintf (string, "%2.2ld", lrint(calendar.sec));
+			(P->date.compact) ? sprintf (string, "%d", (int)lrint(calendar.sec)) : sprintf (string, "%02d", (int)lrint(calendar.sec));
 			break;
 		default:
 			GMT_report (C, GMT_MSG_FATAL, "Error: wrong unit passed to GMT_get_time_label\n");
